Refactor: Move helper classes
