+++ vclib-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidc/helpers/AuthenticationResponsePreparationHelper.kt
@@ -1,6 +1,8 @@
-package at.asitplus.wallet.lib.oidc
+package at.asitplus.wallet.lib.oidc.helpers
 
 import at.asitplus.wallet.lib.data.dif.PresentationPreparationHelper
+import at.asitplus.wallet.lib.oidc.AuthenticationRequestParameters
+import at.asitplus.wallet.lib.oidc.RelyingPartyMetadata
 import kotlinx.serialization.Serializable
 
 @Serializable
