+++ vclib-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidc/helpers/ClientIdSchemeParameters.kt
@@ -1,6 +1,7 @@
-package at.asitplus.wallet.lib.oidc
+package at.asitplus.wallet.lib.oidc.helpers
 
 import at.asitplus.crypto.datatypes.pki.X509Certificate
+import at.asitplus.wallet.lib.oidc.OpenIdConstants
 import kotlinx.serialization.Serializable
 
 @Serializable
